Removed various redundant code.
Removed various redundant empty return statements.
Removed unused (duplicated) time-printing shortcut() function.
Patched various uninitialized variables and signed/unsigned mismatches.

+++ ext/lib/Procframe/CodeKalmanSolver.hpp
@@ -383,7 +383,7 @@ namespace gpstk
 
 
          /// Number of measurements
-      int numMeas;
+      size_t numMeas;
 
 
          /// Pointer to stochastic model for dx (or dLat) coordinate
